refactor: base card
