+++ project-canvas/src/components/StoryMapView/Cards/Add/AddCard.tsx
@@ -2,7 +2,7 @@ import { PaperProps } from "@mantine/core"
 import { IconPlus } from "@tabler/icons"
 import { MouseEventHandler } from "react"
 import { Draggable } from "react-beautiful-dnd"
-import { BaseCard } from "../BaseCard"
+import { BaseCard } from "../Base/BaseCard"
 
 export function AddCard({
   id,
