+++ project-canvas/src/components/StoryMapView/Cards/Base/BaseCard.tsx
